check body

+++ src/controllers/tweet.controller.js
@@ -3,6 +3,9 @@ const tweetService = require('../services/tweet.service')
 async function create(req, res, next) {
     try {
         const { content } = req.body
+        if (!content) {
+            return res.status(400).json({"status": 400, "message": "bad request"})
+        }
         const response = await tweetService.create(req.id, content)
         res.status(response.status).json(response)
     }
@@ -15,6 +18,9 @@ async function create(req, res, next) {
 async function update(req, res, next) {
     try {
         const { id, content } = req.body
+        if (!id && !content) {
+            return res.status(400).json({"status": 400, "message": "bad request"})
+        }
         const response = await tweetService.update(req.id, id, content)
         res.status(response.status).json(response)
     }
@@ -38,6 +44,9 @@ async function get(req, res, next) {
 async function delete_tweet(req, res, next) {
     try {
         const { id } = req.body
+        if (!id) {
+            return res.status(400).json({"status": 400, "message": "bad request"})
+        }
         const response = await tweetService.delete_tweet(req.id, id)
         res.status(response.status).json(response)
     }
